@@ -22,15 +22,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.2.16",
-    "@angular/common": "12.2.16",
-    "@angular/compiler": "12.2.16",
-    "@angular/core": "12.2.16",
-    "@angular/elements": "^12.2.16",
-    "@angular/forms": "12.2.16",
-    "@angular/platform-browser": "12.2.16",
-    "@angular/platform-browser-dynamic": "12.2.16",
-    "@angular/router": "12.2.16",
+    "@angular/animations": "13.3.11",
+    "@angular/common": "13.3.11",
+    "@angular/compiler": "13.3.11",
+    "@angular/core": "13.3.11",
+    "@angular/elements": "^13.3.11",
+    "@angular/forms": "13.3.11",
+    "@angular/platform-browser": "13.3.11",
+    "@angular/platform-browser-dynamic": "13.3.11",
+    "@angular/router": "13.3.11",
     "@project-sunbird/client-services": "^3.4.8",
     "@project-sunbird/sb-styles": "0.0.7",
     "@project-sunbird/sunbird-pdf-player-v9": "^4.8.2",
@@ -44,11 +44,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "12.2.18",
+    "@angular-devkit/build-angular": "13.3.9",
     "@angular-devkit/core": "^14.1.1",
-    "@angular/cli": "12.2.18",
-    "@angular/compiler-cli": "12.2.16",
-    "@angular/language-service": "12.2.16",
+    "@angular/cli": "13.3.9",
+    "@angular/compiler-cli": "13.3.11",
+    "@angular/language-service": "13.3.11",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.20.15",
@@ -63,12 +63,12 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^12.2.7",
+    "ng-packagr": "^13.3.1",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tsickle": "0.39.1",
     "tslint": "~6.1.0",
-    "typescript": "^4.3.5",
+    "typescript": "^4.6.4",
     "watch": "^1.0.2",
     "wrench": "^1.5.9"
   }
